Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> I'm skeptical that the overhead and effort here is worth the benefit.
> Are the current false negatives particularly important in practice?

I just submitted a patch that fixes the issue without incurring any additional overhead (it does not allocate memory).  

I added two tests: one to check that inverted semaphores still work and another to test that the issue I describe is fixed.

I believe the changes are minimal and that they address a serious issue.  The current implementation is over-synchronizing and it is hard to tell whether this over-synchronization is preventing us from finding a serious race condition.

Patch-set: 5