[dev.link] cmd/link: create table for hashing DWARF attributes
Add a new 'attrTab' type the linker, to be used for hashing and
commoning of DWARF attributes.

Motivation: in the existing linker DWARF implementation, we create a
new separate DWAttr object for every attribute on a DWARF DIE, which
is not especially space efficient, since many DIEs have attributes
with the same content. For example, each variable of type 'int' will
have a separate DWAttr object for the DW_AT_type attribute.

With this patch, clients can look up an attribute by value
(attr/cls/value/data), then get an index or token that represents that
attribute. This provides a mechanism for commoning up duplicate
attributes.

This patch introduces the table with its implementation, along with a
few unit tests. A subsequent patch will hook it into the actual linker
DWARF generation.

Change-Id: Ib701608e5ebec970f584877e533d4b1853d24f1f