[dev.link] cmd/link: switch to using hashed DWARF attrs
Change the DWDie type to keep a slice of hashed attribute indices
instead of a linked list of DWAttr objects (should be more memory
efficient), and keep a table in the link context that stores
the contents of the attrs.

Note: as part of this work I have cleaned up a couple of oddities
relating to how attributes are added to DIEs. The old code would allow
you to add an attribute with code X to a die, and then later on add
another attr with the same code, relying on the fact that new
attributes are inserted at the head of the attr list.

Change-Id: Id0d5fd9b089d4f63fb006389721ba7f2a7c5d843