Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> I'm skeptical that the overhead and effort here is worth the benefit.
> Are the current false negatives particularly important in practice?

It's hard to define important here and then it's expensive to measure some definitions of important-ness...
For more global definition of important, somebody at Google could run global presubmit over google3 and assess new races and effects on flakiness of existing races. That's quite expensive. Then it also may be hard to decide on the resulting data, because it will have _some_ effect. But how do we define the bar?
But even if this does not show worthiness, there are also these manual reports about false negatives where users complain that they debugged a bug and it turned out to be a race, but race detector did not detect it. These are then quite expensive to debug on the race detector side, e.g. a very recent one:
https://github.com/golang/go/issues/36794#issuecomment-578889751
You also pointed to some in the past:
https://github.com/golang/go/issues/10589

So if there is already a known source of false negatives and the fix does not increase systematic complexity and there is somebody enthusiastic to work on a fix, I would prefer if we fix it now.

The other case that come up as result of this work -- the one related to chan struct{} -- I would think it's pretty important in practice. chan struct{} is widely used, and turns out currently we massively over-synchronize all involved goroutines.

Patch-set: 5
Reviewer: Gerrit User 5400 <5400@62eb7196-b449-3ce5-99f1-c037f21e1705>