performance improvement for reading go.sum
When reading a `go.sum`，we use `if` to match the last line of the file:
https://github.com/golang/go/blob/e48a83f077e47bd015d4b57e63e9b6fb5e77dc8b/src/cmd/go/internal/modfetch/fetch.go#L378-L383

Given most of the case, a `go.sum` should not be empty. So, we don't need to go through all `if`-`else` for each line.

Change-Id: Ia26c851c1942cfdef1f508eb53f349626ef9c0a9
GitHub-Last-Rev: 8e43c5b7796d25605d10f18911c42133921d611e
GitHub-Pull-Request: golang/go#37578