Update patch set 2
Patch Set 2:

This is essentially splitting the decoder into two steps, which adds quite a lot of complexity - and from the only realistic-looking benchmark we have, CodeUnmarshal, you barely get a 0.50% speed-up.

I personally think this isn't worth the tradeoff. I have looked at this code before, since the addition of good errors has a small but noticeable cost. It's good to write the decoder in a way that its "happy path" is the fastest, but I don't think we should go as far as making the error case unmarshal twice.

Remember that benchmarks like UnmarshalInt64 aren't very representative of real world use cases. Larger ones like Code, with structs and slices, and with more data, are where benchstat numbers are far more important.

I've done quite a lot of work on json performance in the past, but I've been careful to not add two decoder passes, and I think we should keep that property :)

Patch-set: 2