Update patch set 2
Patch Set 2:

Here's an example of how the complexity increases in subtle ways.

What if the variable we unmarshal into implements UnmarshalJSON, and the method behaves differently between the first and second unmarshal passes?

I can think of cases where this could be reasonable. Imagine an UnmarshalJSON whose job is to append to existing data. You could now be doing this incremental operation twice.

Patch-set: 2