runtime: use persistentalloc to allocate buckhash rather than sysAlloc
The size of buckhash isn't necessarily a multiple of page size which
resulted in an error when I was experimenting with increasing the
physical page size. We never free buckhash so it makes sense to use
persistentalloc to allocate it and that way the size gets rounded up
to page size transparently.

Change-Id: I9ebc1ab6301b9632dce31cffb0e0adce0c665e71