runtime: round up the capacity of h.allspans to physical page size
When allocating using sysAlloc the size needs to be a multiple of
the physical page size. The recordspan function wasn't doing this.
This doesn't appear to be a problem when using pre-existing physical
page sizes but was a problem when I experimented with increasing
the physical page size.

Rounding the allocation up to physical page size will stop other
people running into the same issue when experimenting with larger
page sizes.

Change-Id: Ib7e05a29a438e50a32de8786cbdfea1c2d93fb8b