Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > (1 comment)
> 
> I've tried to handle this case as follows:
> 
> if resp, err := altRT.RoundTrip(req); err != ErrSkipAltProtocol {
> 	return resp, err
> }
> if req.outgoingLength() != 0 {
> 	if req.GetBody == nil {
> 		return nil, fmt.Errorf("net/http: unable to retry request")
> 	}
> 	// Rewind the body after ErrSkipAltProtocol, because it can
> 	// be already read (Issue 32441).
> 	// NB: Body should be already closed at this point.
> 	var err error
> 	newReq := *req
> 	newReq.Body, err = req.GetBody()
> 	if err != nil {
> 		return nil, err
> 	}
> 	req = &newReq
> }
> 
> But unfortunately it failed TestTrailersClientToServer_h2, TestNoSniffExpectRequestBody_h2 and bunch of other http2 tests with that new internal error, and I'm not sure how do deal with those tests, sorry.

The `testTrailersClientToServer` helper function needs to be adapted to set a GetBody function on its Request. (It currently uses an `io.MultiReader` in order to inject request mutations. Those mutations are idempotent, so they should be safe to restart, but the `net/http` package has no way to infer that.)

Same for `testNoSniffExpectRequestBody`: the `testErrorReader` is idempotent, but the test function needs to set a `GetBody` method in order to inform `net/http` of that fact.

―

I suspect that the other cases are similar, but if we have to fix up this many tests in `net/http`, it seems likely that users would have to fix up a lot of their own tests too. Per my earlier comment, I think we will need to add a new (optional) method that a RoundTripper can implement to probe the Request without actually consuming it.

If the default RoundTripper implementations provide that method, then I suspect that most of these tests (and therefore most users) will not need to be changed.

Patch-set: 1
Reviewer: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>