[dev.boringcrypto] crypto/internal/boring: make accesses to types with finalizers safer
Some types have a finalizer that will free the underlying C value when
the Go one is garbage collected. It's important that the finalizer
doesn't run while a cgo call is using the undelying C value, so they
require runtime.KeepAlive calls after each use.

This is error prone, so replace it with a closure that provides access
to the underlying C value and then automatically calls KeepAlive.

Change-Id: I6d6f38297cd1cf384a1639974d9739a939cbdbcc