Update patch set 1
Patch Set 1:

If one needs base64 encoding or decoding that's super fast, at the cost of simplicity or security, they can already use third party libraries with unsafe or assembly. The standard library avoids those as much as possible, and I agree with Ian here.

Yes, it will take a while for the compiler to get really good at these optimizations, and that's fine.

Patch-set: 1
CC: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>