Update patch set 1
Patch Set 1:

Sure, I'd be fine with an optimization for the bounds checks. Just make sure to include benchmark numbers to prove that the performance improved; the code has been optimized multiple times in the past, so we can't just work on assumptions.

You should use benchstat with multiple runs, like -count=10. Also see perflock if the variance is too high.

Patch-set: 1