Update patch set 2
Abandoned

Abandoning since the CL was closed. Richard, if you want to investigate other performance improvements in the json package in the future, I'd be happy to bounce ideas. For something that might be controversial or a bit risky, perhaps we could talk about it before you do the work for the CL. I'm reachable via email, gophers.slack.com, and irc.

Patch-set: 2
Status: abandoned
Tag: autogenerated:gerrit:abandon