[dev.link] cmd/link: revise dwarf attrtab map setup
This patch switches the DWARF attr table's internal implementation
from a scheme that hashes the contents of the attr and then looks up
the hash code in a map, to a scheme in which DWAttr objects are used
for the map key.

Former implementation is ugly and complicated (since it has to manage
collisions, etc); latter implementation effectively doubles the
storage needed (since we have to store DWAttr as both key and value),
but to recover some of the performance the table is now being more
selective about which attrs to cache.

Change-Id: I78e68bc512090396ffe86e2990d3b2522fecab80