Update patch set 7
Patch Set 7:

(1 comment)

Before, the dynamic relocations are emitted in order. Now they are not. I'm not really sure address order is required, but at least it is a deterministic order. Now it doesn't look deterministic to me. So this may be a problem for reproducible build.

Do we know how much time are spent in generating dynamic relocations? If it doesn't take much time, I'm not sure we want to go this way.

Patch-set: 7
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>