cmd/go: define a build tag for any GOEXPERIMENT which is enabled.
For each experiment that has been enabled in the toolchain, define a build tag
with the same name which can be used for compiling alternative files for the
experiment. This allows changes for the experiment, like extra struct fields in
the runtime, without affecting the base non-experiment code at all.

I use this capability in my CL for static lock ranking
(https://go-review.googlesource.com/c/go/+/207619), so that static lock ranking
can be fully enabled as a GOEXPERIMENT, but there is no overhead in the runtime
when the experiment is not enabled.

I added a test in cmd/go/testdata/scripts to make sure the build tags are being
defined properly. In order to implement the test, I needed to provide GOEXPERIMENT
to the test scripts (but its value is actually set from objabi.Expstring(), so
that it has the experiments actually baked into the toolchain).

I filed https://github.com/golang/go/issues/37937 to make a builder with
GOEXPERIMENT set to 'staticlockranking', once this change and the static lock
ranking change itself are checked in. The builder will ensure another variant of
GOEXPERIMENT is being tested for this change, as well as checking static
lock ranking in the runtime.

Change-Id: Ieb4b86107238febd105558c1e639d30cfe57ab5c