cmd/go: define a build tag for any GOEXPERIMENT which is enabled.
For each experiment that has been enabled in the toolchain, define a build tag
with the same name which can be used for compiling alternative files for the
experiment. This allows changes for the experiment, like extra struct fields in
the runtime, without affecting the base non-experiment code at all.

I use this capability in my CL for static lock ranking
(https://go-review.googlesource.com/c/go/+/207619), so that static lock ranking
can be fully enabled as a GOEXPERIMENT, but there is no overhead in the runtime
when the experiment is not enabled.

I added a test in cmd/go/testdata/scripts to make sure the build tags are being
defined properly. In order to implement the test, I needed to pass through
GOEXPERIMENT to the test scripts.

Change-Id: Ieb4b86107238febd105558c1e639d30cfe57ab5c