Update patch set 3
Patch Set 3:

PS3 is less unsafe than PS1. The speedup is also smaller. We could get about the same speedup if we could remove the conditional in At2, for which we need to change the representation of external relocs.

Patch-set: 3
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>