[dev.link] cmd/compile, cmd/link: move away from DWARF file renumbering
DO NOT SUBMIT

[Note: still has test failures related to abstract functions].

This patch moves the compiler and linker away from the current scheme
used to generate file references in DWARF subprogram dies.

Up until now the scheme has been to have the compiler emit a special
relocation on a DIE file reference that points to the file symbol in
question. The linker then reads this relocation and updates the addend
to the index of the appropriate file in the line table of
the compilation unit of the DIE (the linker emits the comp unit
file table, so it knows at that point what number use). The drawback
of this scheme is that it requires a lot of relocation processing.

With this patch, we switch to having the compiler emit the file index
directly, and then have the linker use the compiler-generated file table
to emit the line table file section (no renumbering, no relocations, etc).

The one place where this breaks down is with the handling of
abstract functions for DWARF inline subprogram DIEs.  If a function
F in is inlined into more that one compilation unit, there will be
multiple (dupOK) copies of the abstract function DIE for F feeding
into the link, and each of these DIEs comes from a comp unit with
a different file numbering scheme.

The plan is to see how costly it would be to just emit the duplicates
(not try to common them). If the storage expansion is acceptable, then
this seems like it would be a better way to go.

Change-Id: Id4fbe67b28a64200a083e3c5ea358dbe091ec917