Update patch set 1
Patch Set 1:

Thanks, but this is currently actually C code.  If we want to rewrite it into Go code, we would need a curly brace after the else too.  That is, there is no missing close brace; there is no open brace to match the close brace that you added.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>