Update patch set 4
Patch Set 4:

Is this going to be a problem for feature-flagging this when we merge to master? We're planning to copy the whole linker, but I don't think we should do that in the compiler, which means we'd have to keep both object formats working in the compiler for now. (Or did we already drop the old object format from the compiler? I don't remember.)

Patch-set: 4
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>