Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Is this going to be a problem for feature-flagging this when we merge to master? We're planning to copy the whole linker, but I don't think we should do that in the compiler, which means we'd have to keep both object formats working in the compiler for now. (Or did we already drop the old object format from the compiler? I don't remember.)

I have been keeping this in mind. I also don't think we want to copy the whole compiler. We already dropped the old format in the branch. My plan is to add it back in before we merge back. Changes to the compiler have been small so far, and it doesn't sound too bad for me to reintroduce the flag.

Patch-set: 4