Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > Is this going to be a problem for feature-flagging this when we merge to master? We're planning to copy the whole linker, but I don't think we should do that in the compiler, which means we'd have to keep both object formats working in the compiler for now. (Or did we already drop the old object format from the compiler? I don't remember.)
> 
> I have been keeping this in mind. I also don't think we want to copy the whole compiler. We already dropped the old format in the branch. My plan is to add it back in before we merge back. Changes to the compiler have been small so far, and it doesn't sound too bad for me to reintroduce the flag.

Yeah. It's a bit unfortunate that happened long before the feature flag decision, but is still impacted.

My inclination would be to bring back the old format support in the compiler sooner rather than later and start testing that it still works so we're not caught by unexpected surprises when we do merge. The fact that the linker codebase doesn't have to deal with the old format should make that not too bad, I think. As you said, the divergence in the compiler itself isn't that much.

Patch-set: 4
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>