cmd/go/internal/modload: implement the "all" pattern for lazy loading
The new semantics of the "all" package pattern can be implemented
without actually changing module loading per se. This change
implements those semantics, so that the change can be decoupled from
the changes to the module requirement graph.

DO NOT SUBMIT

This code is still failing TestScript/mod_retention, and I haven't
updated the tests to cover the new behavior yet.

For #36460

Change-Id: I0ee8b17afa8b728dc470a42a540fcc01764a4442