Update patch set 3
Patch Set 3:

Hi Cherry, on arm64, we found the linker sets funcAlign to 8 bytes, but the assembler aligns the size of function symbols to 16 bytes, the assembler handling codes "funcAlign = 16, pc += -pc & (funcAlign - 1), c.cursym.Size = pc", so the functions' address is 16 bytes aligned. In order to keep the consistent, do we need to change the funcAlign in the linker to 16 bytes? Thank you.

Patch-set: 3