Update patch set 1
Patch Set 1:

(1 comment)

The thing I worry about is that this only prevents a signal queue from piling up. It does not actually guarantee that the signal target makes forward progress. We could still get interrupted at the same user PC indefinitely, at least theoretically.

Imagine a scenario where an OS prioritizes sending a signal to a thread over running any thread normally. To the extent of even swapping out a normally running thread to swap in a different thread so it can send it a signal. This CL makes sure that there isn't more than 1 signal pending for a particular M, but it still may be the case that we have P pending signals for P (or P-c for some small c, perhaps?) of the Ms. Given some overhead of context switching, I could see the OS using all of its CPU context switching and handling signals, never having spare cycles to actually run not-in-signal-handler code.

I don't know whether OSX, or any other OS, is that aggressive. But I don't know that they aren't, or won't in the future, be like that.

Another worry of mine is that we're returning immediately from preemptM without doing anything, which means unless the caller implements backoff (some do, some don't) we risk burning lots of cpu in a busy loop.

That said, this is quite a bit simpler than my CL. It does fix the examples we have, and we can do something more robust for 1.15.

Patch-set: 1
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>