Update patch set 5
Patch Set 5:

This change is ready for review now.

This is a follow up to CL 223699. It makes os tests pass on windows-amd64-longtest builder on 1.13 release branch.

Patch set 4 is almost identical to patch set 1, which was a clean backport of the original CL 203502. However, it also removes the unused removeAllTestHook hook. I decided to include that because it increases my confidence in that the changes to non-test .go files in this CL are safe to backport.

This uncovers another test failure, this time in cmd/compile/internal/syntax package. Fixing that is out of scope for this CL.

Patch-set: 5
Reviewer: Gerrit User 6005 <6005@62eb7196-b449-3ce5-99f1-c037f21e1705>