cmd/go: make 'go get' and 'go install' avoid updating mtimes of binaries that are not stale
We now treat the binaries generated by 'go build' and 'go test -c' as
always stale (in most cases we already did) to ensure that they are
still updated consistently: builds using Make (which depends on
mtimes) are likely to use 'go build' rather than 'go install', and we
don't want to break them without good reason.

However, binaries in GOBIN, GOROOT/bin, GOPATH/bin, and similar
locations are arguably under the control of the 'go' command and
should be treated as an extension of the build cache: if those
binaries are not stale, they should not be modified, so that commands
like 'go install cmd/cgo' can be used to ensure that they are up to
date without spuriously marking the binaries as modified.

Fixes #37573
Updates #30316

Change-Id: I4adabc61c5a69176ef0ecc2b0ac76323f0de0fbd