[dev.link] cmd/link: add -stats debugging options
DO NOT SUBMIT

[[ Note: this version is specific to the linker's "new" object file
   loader, meaning that the stats include metrics from reading new
   objects, then stats related to the loader (while the loader is still
   alive), and then finally stats on sym.Symbol usage. ]]

Add new top-level option for collecting symbol/section/relocation
statistics. Collects number and size information for symbols read in
from object files, then provides hooks for post-processing the AllSyms
slice in the linker for emitting stats on outstanding symbols.

Also includes a debugging option "-showdead <file>", which emits a
summary of all dead symbols to a file following the deadcode pass.
Useful to find out what's being killed off by deadcode. Note that this
analysis tracks only symbols that have data (len(s.P) != 0). Includes
a breakdown of dead symbols (and their sizes) by symbol kind and by
package. You can also use the "-showdead-details" option, that
produces a listing of every symbol in every package to show what's
live and what's not (be warned that this produces a lot of output), as
well as a "showdead-pkg-relocs=P" option that dumps relocations for
package if in the showdead-details dump. Whew.

For experiments on linker space usage -- not intended to be submitted.

Change-Id: Id823e21c9f62aa6cb829b0dceda6c4311261e1cf