Update patch set 4
Patch Set 4:

(3 comments)

> I am planning to have a builder that enables a GOEXPERIMENT of 'staticlockranking', once my static lock ranking change is in (https://go-review.googlesource.com/c/go/+/207619)

In that case, could you add a `staticlockranking` case to the test, or update the commit message to refer to an issue tracking the new builder and test-update? I'm concerned that the builder coverage for this change right now is effectively zero, so it could regress unnoticed.

Or, to be even more robust, you could structure it more like TestNewReleaseRebuildsStalePackagesInGOPATH (in go_test.go) and copy (or symlink) an entire minimal GOROOT into a new location, in which you can build an actual minimal toolchain with GOEXPERIMENT set.

Patch-set: 4
Reviewer: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>