cmd/api: make NewWatcher populate its own package and import metadata
To reduce the latency penalty, cache the metadata and parallelize
the initial calls to NewWatcher, including in TestMain.

In addition to fixing BenchmarkAll, this eliminates the need for
assumptions about the consistency of 'std' across platforms, and for
hard-coded special cases that violate those assumptions.

Tested manually using 'go test -race -bench=. cmd/api'.

Fixes #37951

Change-Id: Icd537e035e725e1ee7c41d97da5c6651233b927e