Update patch set 2
Patch Set 2:

Thank you for this change Michael. If possible, it would be nice for us to have a test because as per what motivated you to make this change, it can be a subtle regression if we stop panicking on undefined intrinsics. it could be in perhaps another CL even.

Patch-set: 2
CC: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>