Update patch set 1
Patch Set 1: Code-Review+2

Short answer is this CL makes it better, but not all the way better.  I'd call it a +2 for now, and resolve to replace the single instruction compare-laod-indexed with the faster load-indexed + compare pair.

It seems to me we could fix this in a post-pass of the otherwise desirable new address-mode-hacking pass.

I measured a bunch of benchmarks on two different values of Intel.  Laptop was "meh", workstation (that experienced the 70% slowdown) now sees about a 40% slowdown on the worst, 10% on a similar small loop benchmark, and everything else is "meh".

Patch-set: 1
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2